requirements.txt
opencc_python_reimplemented==0.1.7
pdfplumber==0.11.6
pypinyin==0.54.0
